insights_service/requirements.txt
fastapi
langchain_google_genai
pydantic
pymongo
python-dotenv
uvicorn